Move parent liveness watching into core process
Constraint: The parent watcher manages the OpenChrome host process lifecycle, not Chrome child cleanup.\nRejected: Grouping it with Chrome process guardian | that would conflate MCP host liveness with Chrome child process ownership.\nRejected: Keeping parent-watcher under utils | it is runtime lifecycle behavior, not a leaf helper.\nConfidence: high\nScope-risk: narrow\nDirective: Keep host-process lifecycle primitives under src/core/process unless they become transport-specific.\nTested: npx jest --config jest.ci.config.js --runInBand --silent --runTestsByPath tests/core/process/parent-watcher.test.ts tests/integration/embedded-server.test.ts tests/integration/stdio-orphan.test.ts\nTested: npm run build\nTested: npm run lint\nTested: npm run lint:tier\nTested: npm run lint:repo-structure\nTested: git diff --check\nNot-tested: Full Jest suite.
